Replace `FakeHTTP` with httpmock which is now compatible
